Renamed api
