+++ src/sophronia/app/(pages)/time/[year]/q/[quarter]/page.tsx
@@ -1,5 +1,8 @@
 import { TimePage } from 'app/(pages)/time/time';
-import { fetchBoxOfficeInfinite, fetchBoxOfficeSummary } from 'lib/fetch/box';
+import {
+	fetchBoxOfficeInfinite,
+	fetchBoxOfficeSummary,
+} from 'lib/fetch/boxoffice';
 import { getLastDayofMonth } from 'lib/utils/dates';
 
 export async function generateMetadata({
