+++ src/sophronia/app/page.tsx
@@ -1,7 +1,7 @@
 import {
 	fetchBoxOfficeInfinite,
 	fetchBoxOfficePreviousYear,
-} from 'lib/fetch/box';
+} from 'lib/fetch/boxoffice';
 import { groupForTable } from 'lib/utils/groupData';
 import { Controls } from 'components/controls';
 import { FilmTableDetailed } from 'components/tables/film-table-detailed';
